% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centroid.R
\name{calculate_centroid}
\alias{calculate_centroid}
\title{Calculate Centroid}
\usage{
calculate_centroid(dbr, rKey, geom)
}
\description{
Calculates the centroid from a vector of
 longitdues and latitudes and updates the appropriate
 redis key.
}
