% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zoom.R
\name{calculate_zoom}
\alias{calculate_zoom}
\title{Calculate Zoom}
\usage{
calculate_zoom(zoomExtremes)
}
\arguments{
\item{zoomExtremes}{A named list containing
\itemize{
  \item{long: A vector of length two containing the min and max longitudes}
  \item{lat: A vector of length two containing the min and max latitudes}
}}
}
\value{
An appropriate zooming factor to encapsulate all co-ordinates.
}
\description{
A function to calculate the correct zoom
 parameter from a list of longitude and latitude extremeties.
}
