% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data.R
\name{park_garden_data}
\alias{park_garden_data}
\title{Park and Garden Data}
\usage{
park_garden_data(dbr, update = FALSE)
}
\arguments{
\item{dbr}{A redis environment containing all redis
functionality.}
}
\description{
A function to insert park and garden data,
 original data source from \code{https://www.opendatani.gov.uk/dataset/c6529642-8fa0-4205-bd56-734e9fda2cb0/resource/bbb60928-d39b-4660-8bef-280f17e70bd8/download/hpag.geojson}.
}
