% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_closest.R
\name{return_closest}
\alias{return_closest}
\title{Return Closest}
\usage{
return_closest(dbr, allKeys, currentLoc, returnItems)
}
\description{
A function returning the closest by distance
 from a single location.
}
\details{
This function takes a vector of redis keys that
 should contain the following fields in the hash,
 \itemize{
   \item{CENT_LONG: The longitute of the centroid}
   \item{CENT_LAT: The latitude of the centroid}
 }
then returns the keys that are closest.
}
